pyproject.toml
[tool.poetry]
name = "lms-api"
version = "0.1.0"
description = ""
authors = ["Oleksandr Sviatetskyi <alex.sviatetskyi@gmail.com>"]

[tool.poetry.dependencies]
python = "^3.9"
Django = "^3.2"
djangorestframework = "^3.12.4"
dj-rest-auth = "^2.1.4"
drf-spectacular = "^0.15.1"
gunicorn = "^20.1.0"
django-environ = "^0.4.5"
psycopg2-binary = "^2.8.6"
django-filter = "^2.4.0"
django-crispy-forms = "^1.11.2"
rest-framework-generic-relations = "^2.0.0"

[tool.poetry.dev-dependencies]
autopep8 = "^1.5.6"
pylint-django = "^2.4.3"
pylint = "^2.8.1"

[build-system]
requires = ["poetry-core>=1.0.0"]
build-backend = "poetry.core.masonry.api"
